requirements.txt
Django==5.2.8
djangorestframework==3.16.1
python-dotenv==1.2.1
django-storages==1.14.6
boto3==1.41.4
pillow==12.0.0
psycopg2-binary==2.9.11
